add user authorization and authentication

+++ backend/controllers/inventory.controller.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler');
 
 const Inventory = require('../models/inventory.model');
+const User = require('../models/user.model');
 
 // @desc  Create Inventory Items
 // @route  POST /api/inventory
@@ -13,7 +14,8 @@ const createInventory = asyncHandler(async (req, res) => {
     const inventory = await Inventory.create({
         name: req.body.name,
         stock: req.body.stock,
-        price: req.body.price
+        price: req.body.price,
+        user: req.user.id
     })
     res.status(200).json(inventory)
 })
@@ -21,7 +23,7 @@ const createInventory = asyncHandler(async (req, res) => {
 // @desc  Get Inventory Items
 // @route  GET /api/inventory
 const getInventory = asyncHandler(async (req, res) => {
-    const inventory = await Inventory.find()
+    const inventory = await Inventory.find({ user: req.user.id })
 
     res.status(200).json(inventory)
 })
@@ -43,6 +45,22 @@ const updateInventory = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Inventory not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the login user matches the inventory user
+    if(inventory.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+
     const updateInventory = await Inventory.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updateInventory)
 })
@@ -56,6 +74,22 @@ const deleteInventory = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Inventory not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the login user matches the inventory user
+    if(inventory.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    
     const deletedInventory = await Inventory.findByIdAndRemove(req.params.id, req.body)
     res.status(200).json(`Deleted ${inventory}`)
 })
